KC2021.05: Fixes #12
In order to solve the corresponding requirement, two new templates were
added:
 - formLogin ("Hinweise während des Logins")
 - formIntroduction ("Einleitung zur Veranstaltung")

Both new templates can be assigned to the template set.
Furthermore the event is available as context (variable: "event").

If there is nothing configured for formLogin, a default text is shown to
the user.

In case of the event form introduction ("formIntroduction"), the user
does not get any default text shown.