Closes #13: Distinguish between digital and on-site
With this changes, issue #13 is fixed.
A new option is available on event level, called "On-site attendance"
("Präsenzteilnahme"), which enforces further documents and information
from participant.

In case of a digital event, this checkbox can be omited. The user is not
annoyed in entering nutrition or to upload further documents.
Even the event passport is optional in this case.